+++ src/Shop/CalculatorService.php
@@ -2,6 +2,8 @@
 
 namespace DDDHH\Shop;
 
+use DDDHH\Shop\Cart\Cart;
+
 interface CalculatorService
 {
     /**
@@ -10,5 +12,5 @@ interface CalculatorService
      * @param string[] $discountedItemIds
      * @return float
      */
-    public function total(ShoppingCart $cart, float $discount = 1.0, array $discountedItemIds = []): float;
+    public function total(Cart $cart, float $discount = 1.0, array $discountedItemIds = []): float;
 }
