+++ src/Shop/ImperativeCalculatorService.php
@@ -2,6 +2,8 @@
 
 namespace DDDHH\Shop;
 
+use DDDHH\Shop\Cart\Cart;
+
 class ImperativeCalculatorService implements CalculatorService
 {
     /** @var bool */
@@ -13,7 +15,7 @@ class ImperativeCalculatorService implements CalculatorService
      * @param string[] $discountedItemIds
      * @return float
      */
-    public function total(ShoppingCart $cart, float $discount = 1.0, array $discountedItemIds = []): float
+    public function total(Cart $cart, float $discount = 1.0, array $discountedItemIds = []): float
     {
         $total = 0.0;
         foreach ($cart->items() as $item) {
