+++ tests/Shop/Cart/CartTest.php
@@ -1,19 +1,19 @@
 <?php
 
-namespace DDDHH\Shop;
+namespace DDDHH\Shop\Cart;
 
-use DDDHH\Shop\ShoppingCart\Item;
+use DDDHH\Shop\Customer;
 
 use PHPUnit\Framework\TestCase;
 
-class ShoppingCartTest extends TestCase
+class CartTest extends TestCase
 {
     /**
      * @test
      */
     public function itShouldAddRetrieveAndDeleteItem()
     {
-        $cart = new ShoppingCart(new Customer\Id('XXSS-1234'));
+        $cart = new Cart(new Customer\Id('XXSS-1234'));
 
         $item = new Item(
             'AAXX-4711',
@@ -55,7 +55,7 @@ class ShoppingCartTest extends TestCase
             ),
         ];
 
-        $cart = new ShoppingCart(new Customer\Id('XXSS-1234'), $expectedItems);
+        $cart = new Cart(new Customer\Id('XXSS-1234'), $expectedItems);
         $items = $cart->items();
 
         foreach ($expectedItems as $expectedItem) {
@@ -69,7 +69,7 @@ class ShoppingCartTest extends TestCase
     public function itShouldInitializeWithId()
     {
         $customerId = new Customer\Id('BBCC-8342');
-        $cart = new ShoppingCart($customerId);
+        $cart = new Cart($customerId);
 
         $this->assertEquals($customerId, $cart->customerId());
     }
